listening-comp/requirements.txt
streamlit
youtube_transcript_api
google-generativeai
chromadb
gTTS
python-dotenv